fix(mcp): drop unnecessary i64 cast in find_workflow fallback
clippy::unnecessary_cast — `limit` is already i64 (from
FindWorkflowParams.limit: Option<i64>), so `(limit * 2) as i64` was a
no-op. CI clippy was failing with -D warnings.

+++ crates/epigraph-mcp/src/tools/workflows.rs
@@ -291,7 +291,7 @@ pub async fn find_workflow(
             &["workflow".to_string()],
             &params.goal,
             min_truth,
-            (limit * 2) as i64,
+            limit * 2,
         )
         .await
         .unwrap_or_else(|e| {
